new drivers/spi/spi-tty-plain module
Original code was copied from spi-tty-ds.c (thanks to Federico Vaga).
This driver is a simplification and I called it 'plain', since it
needs no protocol headers and no interrupts.

Communication from slave to master is implemented by means of a workqueue
which polls the slave by sending the ascii "enquiry" byte, that is 0x5.

The obtained tty is slow (see #define SPI_TTY_FREQ_HZ 38400bps) but reliable
useful for a login shell.

Signed-off-by: Aurelio Colosimo <aurelio@aureliocolosimo.it>