drivers/pwm/pwm-mcuio.c: implement update_period and update_duty
These two functions are needed because AVR chip shares the same timer
among various pwm outputs: a change to the period of one of them causes
a change in the others as well. Since pwm framework was not aware of
this situation, I patched it and used the new functions.

Signed-off-by: Aurelio Colosimo <aurelio@aureliocolosimo.it>