[REGMAP] regmap-mcuio: use mcuio request fill flag
If a total of more than 8 bytes must be read/written, try using
as much as fill messages as we can, to reduce mcuio protocol overhead
effect (8 bytes of data can be transferred via a single mcuio message,
instead of just 1, 2, 4). This is particularly useful in case
regmap_raw_read/write functions are used.